Add form.media to form
This will probably come in handy...

I've tested this and it was output correctly. And the JS was allowed by the browser.

However, I couldn't use it for anything because Django's AdminSplitDateTime widget is loaded with window.load event, and changing it seemed to require copying the entire JS file.